move bitcoin utils to separate package
